Added support for an idle callback.
To make regular updates to the GUI while monitoring a subprocess or
retrieving data from another thread without freezing the entire application.

Note: For practical/implementation reason, the idle() methods have been added
to the UIElement class, but are not conceptualy bound to it.

Missing: HTML documentation upgrade. The gen-docs.sh script alone does not
generate documents with the proper styling.